requirements.txt
black
flake8
isort
loguru
mkdocs
pip
pytest
python-dotenv
tqdm
typer
pathlib
ipykernel
ipywidgets
numpy
pandas
matplotlib
tensorflow
keras
tensorflow_datasets
duckdb
-e .
